Update to use VideoView from https://github.com/radiolondra/libvlcsharp/tree/3.x/src/LibVLCSharp.Avalonia.Unofficial
This version is meant to be able to work inside a UserControl, by
allowing the handle to be set manually. More works needs to be done to
clean up the VideoView code (as it is not very well written), possibly
removing the requirement for CSharpFunctionalExtensions, and
ensuring it works in a UserControl.