Re-apply changes from my feature-bgvidcap branch for hotspot_detection.py
Tested and works pretty well. Works when hotspot detection is stopped, the hotspots moved around, camera recalibrated, and then the hotspot detection started again. Hotspot detection is a little temperamental at times, but I reckon this is just the mediapipe model and not a regression in my code.

Still to do: re-apply changes for VideoCaptureThread.py